impl-submit-report - move BarkReportData into bark/models
