+++ src/app/vet/(logged-in)/appointments/_components/submit-report-form.tsx
@@ -1,3 +1,4 @@
+import { BarkForm } from "@/components/bark/bark-form";
 import { BarkAppointment } from "@/lib/bark/bark-models";
 
 export function SubmitReportForm(props: { appointment: BarkAppointment }) {
