+++ src/lib/bark/mappers/to-encrypted-bark-report-data.ts
@@ -1,5 +1,6 @@
 import { BarkContext } from "../bark-context";
-import { BarkReportData, EncryptedBarkReportData } from "../bark-models";
+import { EncryptedBarkReportData } from "../bark-models";
+import { BarkReportData } from "../models/bark-report-data";
 
 export async function toEncryptedBarkReportData(
   context: BarkContext,
