+++ tests/bark/operations/op-fetch-report.test.ts
@@ -1,4 +1,5 @@
-import { BarkReport, BarkReportData } from "@/lib/bark/bark-models";
+import { BarkReport } from "@/lib/bark/bark-models";
+import { BarkReportData } from "@/lib/bark/models/bark-report-data";
 import { givenDog, givenVet } from "../_given";
 import { mockReportData } from "../_mocks";
 import { withBarkContext } from "../_context";
